requirements.txt
numpy
matplotlib
texttable
tqdm
seaborn
pyyaml